Add caches_page :index to WelcomeController
Make the site faster by caching the page. This requires moving the
data related to the current_user to the front-end (with JS/ajax) and
temporarily disabling flash messages.

The current_user data is retrieved with a GET to /sessions/index. This
call is only performed if a current user might exist, which is checked
by means of a cookie